Enable Babel loose mode, set deps to babel alpha version
Unfortunately, the loose mode changes are not yet released and NPM does not support installing packages from Github projects where package.json is not in the root. Babel is a monorepo, so there does not seem to be an easy way to test with alpha packages before they are released with a version number on NPM.

By the time this issue is picked up, update the alpha versions in package.json.